Added methods for bin refinement, better bin generation, and cell aggregation

+++ scripts/analyze.sh
@@ -23,6 +23,8 @@ then
   genome=${genome}/pseudoautosomal
 else
   genome=${genome}/original
+  #genome=${genome}/win0.2/p0.6
+  #genome=${genome}/win0.2/p0.5635
 fi
 
 # ------------------------------------------------------------------------------
@@ -45,13 +47,22 @@ fi
 # -- Map Reads & Prepare Samples For Processing
 # ------------------------------------------------------------------------------
 
+timeA=`date +%s.%N`
+
 total=`wc -l < ${dir}/${inFile}`
+
+export init="1";
 
 if [ "$init" == "1" ];
 then
 
-  # Clean directory
-  rm -f ${dir}/*_mapped ${dir}/*.jpeg ${dir}/*.newick ${dir}/*.xml ${dir}/*.cnv ${dir}/Seg* ${dir}/results.txt
+  if [ -z "$clustThresh" ];
+  then
+    echo "Cleaning directory"
+    # Clean directory
+    rm -f ${dir}/*_mapped ${dir}/*.jpeg ${dir}/*.newick ${dir}/*.xml ${dir}/*.cnv ${dir}/Seg* ${dir}/results.txt
+    rm -r -f ${dir}/CNVprofiles
+  fi
 
   # Map user bed files to appropriate bins
   cnt=0
@@ -69,7 +80,24 @@ then
         mv ${dir}/${file}_tmp ${dir}/${file/.gz/}
         gzip -f ${dir}/${file/.gz/}
       fi
-      ${home}/scripts/binUnsorted ${genome}/${binMeth} `wc -l < ${genome}/${binMeth}` <(zcat -cd ${dir}/${file}) `echo ${file} | awk -F ".bed" '{print $1}'` ${dir}/${file}_mapped
+
+      if [ -f ${dir}/${file}.cells ];
+      then
+        echo "${home}/scripts/binUnsorted ${genome}/${binMeth} `wc -l < ${genome}/${binMeth}` <(zcat -cd ${dir}/${file}) `echo ${file} | awk -F ".bed" '{print $1}'` ${dir}/${file}_mapped_ `wc -l < ${dir}/${file}.cells` ${dir}/${file}.cells"
+        ${home}/scripts/binUnsorted ${genome}/${binMeth} `wc -l < ${genome}/${binMeth}` <(zcat -cd ${dir}/${file}) `echo ${file} | awk -F ".bed" '{print $1}'` ${dir}/${file}_mapped_ `wc -l < ${dir}/${file}.cells` ${dir}/${file}.cells
+      else
+        echo "${home}/scripts/binUnsorted ${genome}/${binMeth} `wc -l < ${genome}/${binMeth}` <(zcat -cd ${dir}/${file}) `echo ${file} | awk -F ".bed" '{print $1}'` ${dir}/${file}_mapped 1"
+        ${home}/scripts/binUnsorted ${genome}/${binMeth} `wc -l < ${genome}/${binMeth}` <(zcat -cd ${dir}/${file}) `echo ${file} | awk -F ".bed" '{print $1}'` ${dir}/${file}_mapped 1
+      fi
+
+      if [ "$improvebounds" == "1" ]; then
+        if [ -f ${dir}/${file}.cells ];
+        then
+          ${home}/scripts/binUnsorted ${genome}/${binMethFine} `wc -l < ${genome}/${binMethFine}` <(zcat -cd ${dir}/${file}) `echo ${file} | awk -F ".bed" '{print $1}'` ${dir}/${file}_mappedfine_ `wc -l < ${dir}/${file}.cells` ${dir}/${file}.cells
+        else
+          ${home}/scripts/binUnsorted ${genome}/${binMethFine} `wc -l < ${genome}/${binMethFine}` <(zcat -cd ${dir}/${file}) `echo ${file} | awk -F ".bed" '{print $1}'` ${dir}/${file}_mappedfine 1
+        fi
+      fi
 
     # 
     else
@@ -79,7 +107,22 @@ then
         awk '{print "chr"$0}' ${dir}/${file} > ${dir}/${file}_tmp
         mv ${dir}/${file}_tmp ${dir}/${file}
       fi
-      ${home}/scripts/binUnsorted ${genome}/${binMeth} `wc -l < ${genome}/${binMeth}` ${dir}/${file} `echo ${file} | awk -F ".bed" '{print $1}'` ${dir}/${file}_mapped
+
+      if [ -f ${dir}/${file}.cells ];
+      then
+        ${home}/scripts/binUnsorted ${genome}/${binMeth} `wc -l < ${genome}/${binMeth}` ${dir}/${file} `echo ${file} | awk -F ".bed" '{print $1}'` ${dir}/${file}_mapped_ `wc -l < ${dir}/${file}.cells` ${dir}/${file}.cells
+      else
+        ${home}/scripts/binUnsorted ${genome}/${binMeth} `wc -l < ${genome}/${binMeth}` ${dir}/${file} `echo ${file} | awk -F ".bed" '{print $1}'` ${dir}/${file}_mapped 1
+      fi
+
+      if [ "$improvebounds" == "1" ]; then
+        if [ -f ${dir}/${file}.cells ];
+        then
+          ${home}/scripts/binUnsorted ${genome}/${binMethFine} `wc -l < ${genome}/${binMethFine}` ${dir}/${file} `echo ${file} | awk -F ".bed" '{print $1}'` ${dir}/${file}_mappedfine_ `wc -l < ${dir}/${file}.cells` ${dir}/${file}.cells
+        else
+          ${home}/scripts/binUnsorted ${genome}/${binMethFine} `wc -l < ${genome}/${binMethFine}` ${dir}/${file} `echo ${file} | awk -F ".bed" '{print $1}'` ${dir}/${file}_mappedfine 1
+        fi
+      fi
       gzip ${dir}/${file}
     fi
 
@@ -87,9 +130,12 @@ then
   done < ${dir}/${inFile}
 
   # Concatenate binned reads to central file  
-  paste ${dir}/*_mapped > ${dir}/data
-  rm -f ${dir}/*_mapped ${dir}/*_binned
-
+  if [ "$improvebounds" == "1" ]; then
+    paste ${dir}/*_mappedfine* > ${dir}/dataFine
+    rm -f ${dir}/*_mappedfine* ${dir}/*_binned*
+  fi
+  paste ${dir}/*_mapped* > ${dir}/data
+  #rm -f ${dir}/*_mapped* ${dir}/*_binned
 fi
 
 # ------------------------------------------------------------------------------
@@ -97,20 +143,41 @@ fi
 # ------------------------------------------------------------------------------
 
 if [ "$segMeth" == "2" ]; then
-    ${home}/scripts/binUnsorted ${genome}/${binMeth} `wc -l < ${genome}/${binMeth}` ${dir}/${ref} Reference ${dir}/${ref}_mapped
+  if [ -f ${dir}/${ref}.cells ];
+  then
+    ${home}/scripts/binUnsorted ${genome}/${binMeth} `wc -l < ${genome}/${binMeth}` ${dir}/${ref} Reference ${dir}/${ref}_mapped_ `wc -l < ${dir}/${ref}.cells` ${dir}/${ref}.cells
+  else
+    ${home}/scripts/binUnsorted ${genome}/${binMeth} `wc -l < ${genome}/${binMeth}` ${dir}/${ref} Reference ${dir}/${ref}_mapped 1
+  fi
 else
     ref=refDummy.bed
     touch ${dir}/${ref}_mapped
 fi
+
+timeB=`date +%s.%N`
+runtime=$(echo "$timeB-$timeA" | bc)
+rm ${dir}/timing.txt
+echo -e $runtime >> ${dir}/timing.txt
+#echo -e  "Time to map reads: " $runtime >> ${dir}/timing.txt
 
 # ------------------------------------------------------------------------------
 # -- Run Mapped Data Through Primary Pipeline
 # ------------------------------------------------------------------------------
 
 if [ "$process" == "1" ]; then
-  echo "Launching process.R $genome $dir $statFile data $segMeth $binMeth $clustMeth $distMet $color ${ref}_mapped $f $facs $sex $rmbadbins $maxploidy $minbinwidth"
-  ${home}/scripts/process.R $genome $dir $statFile data $segMeth $binMeth $clustMeth $distMet $color ${ref}_mapped $f $facs $sex $rmbadbins $maxploidy $minbinwidth
+  if [ "$improvebounds" == "1" ]; then
+    echo "Launching process.R $genome $dir $statFile data $segMeth $binMeth $clustMeth $distMet $color ${ref}_mapped $f $facs $sex $rmbadbins $maxploidy $minbinwidth $improvebounds $binMethFine dataFine $clustThresh"
+    ${home}/scripts/process.R $genome $dir $statFile data $segMeth $binMeth $clustMeth $distMet $color ${ref}_mapped $f $facs $sex $rmbadbins $maxploidy $minbinwidth $improvebounds $binMethFine dataFine $clustThresh
+  else
+    echo "Launching process.R $genome $dir $statFile data $segMeth $binMeth $clustMeth $distMet $color ${ref}_mapped $f $facs $sex $rmbadbins $maxploidy $minbinwidth $improvebounds $clustThresh"
+    ${home}/scripts/process.R $genome $dir $statFile data $segMeth $binMeth $clustMeth $distMet $color ${ref}_mapped $f $facs $sex $rmbadbins $maxploidy $minbinwidth $improvebounds $clustThresh
+  fi
 fi
+
+timeC=`date +%s.%N`
+runtime=$(echo "$timeC-$timeB" | bc)
+echo -e $runtime >> ${dir}/timing.txt
+#echo -e  "Time to segment:   " $runtime >> ${dir}/timing.txt
 
 # ------------------------------------------------------------------------------
 # -- Recreate Clusters/Heat Maps (With New Parameters)
@@ -121,6 +188,11 @@ if [ "$fix" == "1" ]; then
   ${home}/scripts/reclust.R $genome $dir $statFile $binMeth $clustMeth $distMet $f $facs $sex
 fi
 
+timeD=`date +%s.%N`
+runtime=$(echo "$timeD-$timeC" | bc)
+echo -e $runtime >> ${dir}/timing.txt
+#echo -e  "Time to recluster: " $runtime >> ${dir}/timing.txt
+
 # ------------------------------------------------------------------------------
 # -- Create CNV profiles
 # ------------------------------------------------------------------------------
@@ -128,12 +200,26 @@ fi
 nbCols=$(awk '{ print NF; exit; }' $dir/SegCopy)
 for (( i=1; i<=$nbCols; i++ ));
 do
-  currCell=$(cut -f$i $dir/SegCopy | head -n 1 | tr -d '"')
-  if [ "$currCell" == "" ]; then
-    continue;
+  if [ -z "$clustThresh" ];
+  then
+    currCell=$(cut -f$i $dir/SegCopy | head -n 1 | tr -d '"')
+    if [ "$currCell" == "" ]; then
+      continue;
+    fi
+    cut -f$i $dir/SegCopy | tail -n+2 | awk '{if(NR==1) print "1,"$1; else print NR","prev"\n"NR","$1;prev=$1; }' > $dir/$currCell.cnv
+  else
+    currCell=$(cut -f$i $dir/SegCopyClust | head -n 1 | tr -d '"')
+    if [ "$currCell" == "" ]; then
+      continue;
+    fi
+    cut -f$i $dir/SegCopyClust | tail -n+2 | awk '{if(NR==1) print "1,"$1; else print NR","prev"\n"NR","$1;prev=$1; }' > $dir/$currCell.cnv
   fi
-  cut -f$i $dir/SegCopy | tail -n+2 | awk '{if(NR==1) print "1,"$1; else print NR","prev"\n"NR","$1;prev=$1; }' > $dir/$currCell.cnv
 done
+
+timeE=`date +%s.%N`
+runtime=$(echo "$timeE-$timeD" | bc)
+echo -e $runtime >> ${dir}/timing.txt
+#echo -e  "Time to profile:   " $runtime >> ${dir}/timing.txt
 
 # ------------------------------------------------------------------------------
 # -- Call CNVs
@@ -141,6 +227,16 @@ done
 
 echo "Launching ${home}/scripts/CNVcaller ${dir}/SegCopy ${dir}/CNV1 ${dir}/CNV2"
 ${home}/scripts/CNVcaller ${dir}/SegCopy ${dir}/CNV1 ${dir}/CNV2
+
+mkdir -p ${dir}/CNVprofiles
+cp ${dir}/*_CN.jpeg ${dir}/CNVprofiles/
+tar -zcvf ${dir}/CNVprofiles.tar.gz ${dir}/CNVprofiles
+
+timeF=`date +%s.%N`
+runtime=$(echo "$timeF-$timeE" | bc)
+echo -e $runtime >> ${dir}/timing.txt
+echo -e "Done" >> ${dir}/timing.txt
+#echo -e  "Time to call CNVs: " $runtime >> ${dir}/timing.txt
 
 # ------------------------------------------------------------------------------
 # -- Email notification of completion
